Make movement keyboard based and simplify code
