+++ entities/spaceship/Spaceship.cs
@@ -2,91 +2,50 @@ using Godot;
 
 public partial class Spaceship : CharacterBody3D
 {
-    public const float MAX_SPEED = 10.0f;
-    public const float THRUSTER_FORCE = .2f;
+    public const float MAX_SPEED = 50.0f;
+    public const float THRUSTER_FORCE = .5f;
     private Vector3? _previousPosition;
-    private int _movementOptionSelected = 0;
-
-
-    // Get the gravity from the project settings to be synced with RigidBody nodes.
-    public float gravity = ProjectSettings.GetSetting("physics/3d/default_gravity").AsSingle();
 
     /// <inheritdoc />
     public override void _PhysicsProcess(double delta)
     {
         DoRotate();
-        DoMove();
+        DoMove((float)delta);
         DoCameraFollow();
     }
 
     private void DoRotate()
     {
-        if (_movementOptionSelected is not 3)
-        {
-            // Get global mouse position.
-            Vector2 mousePos = GetViewport().GetMousePosition();
-
-            Vector3 rayOrigin = GetViewport().GetCamera3d().ProjectRayOrigin(mousePos);
-            Vector3 rayDirection = GetViewport().GetCamera3d().ProjectRayNormal(mousePos);
-
-            float distance = -rayOrigin.y / rayDirection.y;
-            Vector3 position = rayOrigin + rayDirection * distance;
-
-            LookAt(position);
-        }
-        else
-        {
-            float input = Input.GetAxis("move_left", "move_right");
-            float steerValue = input * 2;
-            RotateY(Mathf.DegToRad(steerValue));
-        }
+        float input = Input.GetAxis("turn_left", "turn_right");
+        float steerValue = input * 2;
+        RotateY(Mathf.DegToRad(steerValue));
     }
 
-    private void DoMove()
+    private void DoMove(float delta)
     {
-        Vector3 velocity = Velocity;
+        Vector2 inputDir = Input.GetVector("move_left", "move_right", "move_forward", "move_back");
+        Vector3 direction = (Transform.basis * new Vector3(inputDir.x, 0, inputDir.y)).Normalized();
 
-        // Get the input direction and handle the movement/deceleration.
-        Vector2 inputDir = Input.GetVector("move_left", "move_right",
-            _movementOptionSelected != 2 ? "move_forward" : "move_forward_mouse",
-            _movementOptionSelected != 2 ? "move_back" : "move_back_mouse");
-
-        Vector3 direction = new Vector3();
-
-        if (_movementOptionSelected is 0)
+        // Apply input force.
+        if (inputDir != Vector2.Zero)
         {
-            direction = new Vector3(inputDir.x, 0, inputDir.y).Normalized();
-        }
-        else if (_movementOptionSelected is 1)
-        {
-            direction = (Transform.basis * new Vector3(inputDir.x, 0, inputDir.y)).Normalized();
-        }
-        else if (_movementOptionSelected is 2 or 3)
-        {
-            direction = (Transform.basis.z * inputDir.y).Normalized();
+            Velocity = Velocity.MoveToward(direction * MAX_SPEED, THRUSTER_FORCE);
         }
 
-        if (direction.x != 0)
-        {
-            // Apply thruster force on the X axis.
-            velocity.x = Mathf.MoveToward(velocity.x, direction.x * MAX_SPEED, THRUSTER_FORCE);
-        }
-
-        if (direction.z != 0)
-        {
-            // Apply thruster force on the Z axis.
-            velocity.z = Mathf.MoveToward(velocity.z, direction.z * MAX_SPEED, THRUSTER_FORCE);
-        }
-
+        // Apply breaking force.
         if (Input.IsKeyPressed(Key.Space))
         {
-            // Apply breaking.
-            velocity.x = Mathf.MoveToward(velocity.x, 0, THRUSTER_FORCE);
-            velocity.z = Mathf.MoveToward(velocity.z, 0, THRUSTER_FORCE);
+            Velocity = Velocity.MoveToward(Vector3.Zero, THRUSTER_FORCE);
         }
 
-        Velocity = velocity;
-        MoveAndSlide();
+        // Move and detect collisions.
+        KinematicCollision3D collisionInfo = MoveAndCollide(Velocity * delta);
+
+        // Apply bounce and friction when colliding with something.
+        if (collisionInfo != null)
+        {
+            Velocity = Velocity.Bounce(collisionInfo.GetNormal()) / 4;
+        }
     }
 
     private void DoCameraFollow()
@@ -100,10 +59,5 @@ public partial class Spaceship : CharacterBody3D
         }
 
         _previousPosition = Position;
-    }
-
-    public void _on_option_button_item_selected(int index)
-    {
-        _movementOptionSelected = index;
     }
 }
